\name{GSAFisherComb-package}
\alias{GSAFisherComb-package}
\alias{GSAFisherComb}
\docType{package}
\title{
\packageTitle{GSAFisherComb}
}
\description{
\packageDescription{GSAFisherComb}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{GSAFisherComb}
\packageIndices{
\code{\link{GSAfisher}}
\code{\link{hist.GSAfisher.mult}}
\code{\link{loadGSAfisher}}
\code{\link{plot.GSAfisher.mult}}
\code{\link{summary.GSAfisher}}
}
}
\author{
\packageAuthor{GSAFisherComb}
Maintainer: \packageMaintainer{GSAFisherComb}
}


