\name{print.GSAfisher}
\alias{print.GSAfisher}
\alias{print.GSAfisher.mult}
\title{
Displays the information in a GSAfisher or GSAfisher.mult object.
}
\description{
If the imput is a GSAfisher object displays the number of pvalues in the object. If it is a GSAfisher.mult object displays the input. 
}
\usage{
\method{print}{GSAfisher}(x,...)
\method{print}{GSAfisher.mult}(x,...)
}
\arguments{
  \item{x}{
A GSAfisher or GSAfisher.mult object.}
  \item{...}{
further arguments passed to or from other methods.
}
}

