\name{pvales.matrix}
\alias{pvales.matrix}
\docType{data}
\title{
Matrix to test GSAfisher.GSAfisher.mult.
}
\description{
Matrix of 5000 columns and 1000 rows.
}
\usage{data("pvales.matrix")}
\format{
  The format is:
 num [1:1000, 1:5000] 0.135 0.133 0.147 0.247 0.25 ...
}
\details{
Each column contains 1000 random values between 0 and 1. Simulating a list of pvalues
}
\examples{
data(pvales.matrix)
}

