\name{pvales.mult}
\alias{pvales.mult}
\docType{data}
\title{
List of vectors to test GSAfisher.GSAfisher.mult.
}
\description{
List of 5000 vectors of 1000 length
}
\usage{data("pvales.mult")}
\format{
  The format is:
List of 5000
 $ : num [1:1000] 0.1102 0.0061 0.1344 0.1744 0.1616 ...
 $ : num [1:1000] 0.0308 0.1227 0.1133 0.1795 0.081 ...
  etc...
}
\details{
Each vector contains 1000 random values between 0 and 1. Simulating a list of pvalues.
}

\examples{
data(pvales.mult)
}

