\name{pvales.totest}
\alias{pvales.totest}
\docType{data}
\title{
Vector of pvalues.
}
\description{
Vector to test the function GSAfisher.
}
\usage{data("pvales.totest")}
\format{
  The format is:
 num [1:1000] 0.235 0.039 0.177 0.204 0.148 ...
}
\details{
1000 values created at random between 0 and 1.
}
\examples{
data(pvales.totest)
}

